Remove nan check in server page test, move FiftyOneJSONEncoder to server json_util module
The nan check in the page test is an incorrect way test NaN/Inifinity serialization now that a
proper JSON encoder has been introduced. JSON encoder tests will be added in a forthcoming commit.
