+++ fiftyone/server/util.py
@@ -15,7 +15,6 @@ import os
 import io
 import struct
 
-
 FILE_UNKNOWN = "Sorry, don't know how to get size for this file."
